refactor(test-support): split panic hook helpers into their own module
Mirrors the existing async_utils/trace_recorder submodule split and the
tests/common/panic_hook.rs counterpart, instead of leaving panic-hook
code as the one exception living directly in test_support.rs.
